Step 1: Make all rendering use anchors
This required use of sprites as borders don't yet implement
anchors. Otherwise, everything now _looks_ the same as it did
before (including a slight bug when facing left and slashing, Not-Link
moves backwards rather than foreward into the slash).

I thnk this is already an improvement! All the offets are now relative
to the center which means they can be symmetrical, which will allow us
to simplify a couple of things even further in the next commit.

I also think things are a bit easier to reason about, eg.:
The 9.25 offset for the slash animation is a stylistic choice to have
Not-Link lean into the hit. 41.25 was that same 9.25 plus 32, which is
half of the extra width around each tile compared to the standing tile.