Step 3: Fix a shared bug
In the original version, fixing this is much more complicated!

I was mistaken in my previous commit comment that 41.25 was 32 +
9.25. The scaling of the sprite is totally different, so it's
actually ((165 - 64) / 2) - 9.25.

So to fix the same issue you'd have to do something like this

`x: player.x - 50.5 + player.dir_x.sign * 9.25,`

It took me much longer to work that out so I could write this
explanation than it did to just fix the issue in my new version!