Step 4: Fix an uncovered bug
I've been assuming I can rely on dir_x to be one way or the other, but
that's not currently true as it can get reset to 0 which since the
refactor, breaks the calculation for the hitbox when moving straight
up/down.

In the original sample this leads to a bit of weirdness where if you
move straight up and down you always face left, and this fixes that
issue as well.