Step 5: Fix an introduced bug
After all this, we're still left with the fact that intersect_rect?
doesn't take into account our anchors, so we need to convert our rects
accordingly.

It's not a hard thing to do, but it would be nicer not to have to
remember to do this! I hope that this repo demonstrates that it is
desirable to specify hit-boxes using the same anchor as the main
sprite, and as such it would make sense for intersect_rect? to treat
rects the same way that the rendering does.